Convert original metadata table code into three components
