+++ src/components/Tables/ClinMetadataTable.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import {
   useTable, useSortBy, usePagination, useFilters,
@@ -6,38 +6,19 @@ import {
 } from 'react-table';
 
 // reactstrap components
-import {
-  Card, Row, Button, Container, InputGroup, InputGroupAddon,
-  Input, InputGroupText, InputGroupButtonDropdown, DropdownToggle,
-  DropdownMenu, DropdownItem, Collapse,
-} from 'reactstrap';
+import { Card, Row } from 'reactstrap';
+import Styles from 'assets/css/StyledComponents/MetadataTableStyled';
+import { DefaultColumnFilter, FuzzyTextFilterFn } from 'components/Filters/filters';
 
-import BASE_URL from '../constants/constants';
-import { ClinMetadataDropdown } from '../components/Dropdown/ClinMetadataDropdown';
-import Styles from '../assets/css/StyledComponents/MetadataTableStyled';
-import {
-  GlobalFilter, DefaultColumnFilter, FuzzyTextFilterFn,
-} from '../components/Filters/filters';
-import {PaginationBar} from  '../components/Tables/Pagination'
-import PATIENT from 'constants/patients_local';
+import {PaginationBar} from  'components/Tables/Pagination'
+import {DataControl} from 'components/Tables/DataControls'
 
-const IndeterminateButton = React.forwardRef(
-  ({ indeterminate, ...rest }, ref) => {
-    const defaultRef = React.useRef();
-    const resolvedRef = ref || defaultRef;
-
-    React.useEffect(() => {
-      resolvedRef.current.indeterminate = indeterminate;
-    }, [resolvedRef, indeterminate]);
-
-    return <Button ref={resolvedRef} {...rest} />;
-  },
-);
 
 FuzzyTextFilterFn.autoRemove = (val) => !val;
 
 
-function Table({ columns, data, metadataCallback }) {
+
+export function ClinMetadataTable({ columns, data, metadataCallback }) {
   const filterTypes = React.useMemo(
     () => ({
       // Add a new fuzzyTextFilterFn filter type.
@@ -79,52 +60,17 @@ function Table({ columns, data, metadataCallback }) {
   useFilters, useGlobalFilter, useGroupBy,
   useSortBy, useExpanded, usePagination);
 
-  const [dropdownOpen, setDropdownOpen] = useState(false);
-  const toggleDropDown = () => setDropdownOpen(!dropdownOpen);
-  const [isCollapsed, setIsCollapsed] = useState(false);
-  const toggleIsCollapsed = () => setIsCollapsed(!isCollapsed);
 
   const pagination = {canPreviousPage, canNextPage, pageOptions, pageCount,
     gotoPage, nextPage, previousPage, setPageSize, state: { pageIndex, pageSize },
   }
+  const topBarFxns = {metadataCallback, toggleHideAllColumns, preGlobalFilteredRows, 
+    setGlobalFilter, state, allColumns
+  }
 
   return (
     <>
-      <Row>
-        <InputGroup>
-          <InputGroupAddon className="tableControls" addonType="prepend">
-            <ClinMetadataDropdown className="dataDropdown" metadataCallback={metadataCallback} />
-            <IndeterminateButton className="toggleAll" onClick={() => toggleHideAllColumns()}> Toggle all </IndeterminateButton>
-            <Button className="toggleColumn" color="primary" onClick={toggleIsCollapsed}> Column Toggles </Button>
-            <InputGroupText className="globalSearchText">Search</InputGroupText>
-          </InputGroupAddon>
-          <GlobalFilter
-            className="globalSearchBar"
-            preGlobalFilteredRows={preGlobalFilteredRows}
-            globalFilter={state.globalFilter}
-            setGlobalFilter={setGlobalFilter}
-          />
-        </InputGroup>
-      </Row>
-      <Row>
-        <Collapse isOpen={isCollapsed}>
-          <Card>
-            <Container>
-              {allColumns.map((column) => (
-                <Button
-                  onClick={() => column.toggleHidden()}
-                  key={column.id}
-                  active={column.isVisible}
-                >
-                  {' '}
-                  {column.id}
-                  {' '}
-                </Button>
-              ))}
-            </Container>
-          </Card>
-        </Collapse>
-      </Row>
+      <DataControl topBarFxns={topBarFxns}/>
       <Row>
         <Card>
           <Styles>
@@ -205,7 +151,7 @@ function Table({ columns, data, metadataCallback }) {
                 })}
               </tbody>
             </table>
-            <PaginationBar paginationFxns={pagination}></PaginationBar>
+            <PaginationBar paginationFxns={pagination}/>
 
           </Styles>
         </Card>
@@ -214,94 +160,13 @@ function Table({ columns, data, metadataCallback }) {
   );
 }
 
-Table.propTypes = {
+ClinMetadataTable.propTypes = {
   columns: PropTypes.arrayOf(PropTypes.object),
   data: PropTypes.arrayOf(PropTypes.object),
   metadataCallback: PropTypes.func,
 };
-Table.defaultProps = {
+ClinMetadataTable.defaultProps = {
   columns: [],
   data: [],
   metadataCallback: () => {},
 };
-
-function CreateColumns(columnNames, cb) {
-  const columnList = [];
-  for (const item in columnNames) {
-    const name = columnNames[item];
-    const column = {
-      Header: (name.charAt(0).toLocaleUpperCase() + name.slice(1)),
-      accessor: name,
-      filter: 'fuzzyText',
-      aggregate: 'count',
-      Aggregated: ({ value }) => `${value} `,
-    };
-    columnList.push(column);
-  }
-  cb(columnList);
-}
-
-function getMetadataData(datasetId, metadata, cb) {
-  const datasets = [];
-  return fetch(`${BASE_URL}/${metadata}/search`, {
-    method: 'POST',
-    body: JSON.stringify({ datasetId }),
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  })
-    .then((response) => {
-      if (response.ok) {
-        return response.json();
-      }
-    })
-    .then((data) => {
-      for (let i = 0; i < data.results[metadata].length; i++) {
-        datasets.push(data.results[metadata][i]);
-      }
-      cb(datasets);
-    });
-}
-
-function TableApp({ datasetId }) {
-  const [selectedMetadata, setSelectedMetadata] = useState('patients');
-  const [data, setData] = useState([]);
-  const [columns, setColumns] = useState([]);
-
-  React.useEffect(() => {
-    // fetch data
-    try {
-      getMetadataData(datasetId, selectedMetadata, setData);
-    } catch (err) {
-      console.log(err);
-    }
-  }, [selectedMetadata, datasetId]);
-
-  React.useEffect(() => {
-    // Separate Effect since state change is async and columns depends on data
-    // Not entirely sure if necessary
-    try {
-      CreateColumns(Object.keys(data[0]), setColumns);
-    } catch (err) {
-      console.log(err);
-    }
-  }, [data]);
-
-  const dataM = React.useMemo(() => data, [data]);
-  const columnsM = React.useMemo(() => columns, [columns]);
-
-  return (
-    <div className="content">
-      <Table columns={columnsM} data={dataM} metadataCallback={setSelectedMetadata} />
-    </div>
-  );
-}
-
-TableApp.propTypes = {
-  datasetId: PropTypes.string,
-};
-TableApp.defaultProps = {
-  datasetId: 'patients',
-};
-
-export default TableApp;
